+++ home.admin/_bootstrap.sh
@@ -1275,6 +1275,7 @@ else
   ############################
 
   echo "# NORMAL START BOOTSTRAP" >> $logFile
+  source <(/home/admin/config.scripts/blitz.data.sh status)
 
   #################################
   # FIX BLOCKCHAINDATA OWNER (just in case)
@@ -1302,6 +1303,22 @@ else
   rm /mnt/hdd/app-data/lnd/logs/${network}/${chain}net/lnd.log 2>/dev/null
   # https://github.com/rootzoll/raspiblitz/issues/1700
   rm /mnt/storage/app-storage/electrs/db/mainnet/LOCK 2>/dev/null
+
+  ####################################
+  # EXPANDING PARTITIONS (for Proxmox)
+  ####################################
+  if [ ${#storagePartition} -gt 0 ] && [ ${#storageUnusedSpacePercent} -gt 0 ] && [ ${storageUnusedSpacePercent} != "0" ]; then
+    echo "# EXPANDING STORAGE PARTITION" >> $logFile
+    /home/admin/config.scripts/blitz.data.sh expand ${storagePartition} >> ${logFile}
+  fi
+  if [ ${#systemPartition} -gt 0 ] && [ "${bootFromStorage}" = "0" ] && [ ${#systemUnusedSpacePercent} -gt 0 ] && [ ${systemUnusedSpacePercent} != "0" ]; then
+    echo "# EXPANDING SYSTEM PARTITION" >> $logFile
+    /home/admin/config.scripts/blitz.data.sh expand ${systemPartition} >> ${logFile}
+  fi
+  if [ ${#dataPartition} -gt 0 ] && [ "${combinedDataStorage}" = "0" ] && [ ${#dataUnusedSpacePercent} -gt 0 ] && [ ${dataUnusedSpacePercent} != "0" ]; then
+    echo "# EXPANDING DATA PARTITION" >> $logFile
+    /home/admin/config.scripts/blitz.data.sh expand ${dataPartition} >> ${logFile}
+  fi
 
 fi