+++ CHANGES.md
@@ -6,7 +6,7 @@ Features:
 
 - New: Boot from NVMe on RaspberryPi5
 - New: Copy from old HDD/SSD/NVMe
-- New: Optional seperation of Data & Storage/Blockchain drives
+- New: Optional seperation of Data & Storage/Blockchain drives with Auto-Expand for Proxmox
 - Update: RaspberryOS arm64 base image 2024-11-19 (Debian 12 Bookworm) [details](https://downloads.raspberrypi.com/raspios_full_arm64/release_notes.txt)
 - Update: amd64 base image: debian-12.8.0-amd64-netinst.iso [details](https://www.debian.org/releases/stable/amd64/release-notes/ch-whats-new.en.html)
 - Update: LNbits 1.0.0 [details](https://github.com/lnbits/lnbits/releases/tag/v1.0.0)